Refactor the gen4 planner
Focused the logic around planning routes to route_planning.go,
and moved everything else to a new file
