+++ go/vt/vtgate/planbuilder/route_planning.go
@@ -33,143 +33,6 @@ import (
 	"vitess.io/vitess/go/vt/vtgate/engine"
 )
 
-var _ selectPlanner = gen4Planner
-
-func gen4Planner(_ string) func(sqlparser.Statement, *sqlparser.ReservedVars, ContextVSchema) (engine.Primitive, error) {
-	return func(stmt sqlparser.Statement, reservedVars *sqlparser.ReservedVars, vschema ContextVSchema) (engine.Primitive, error) {
-		sel, ok := stmt.(*sqlparser.Select)
-		if !ok {
-			return nil, vterrors.Errorf(vtrpcpb.Code_UNIMPLEMENTED, "%T not yet supported", stmt)
-		}
-
-		// handle dual table for processing at vtgate.
-		p, err := handleDualSelects(sel, vschema)
-		if err != nil {
-			return nil, err
-		}
-		if p != nil {
-			return p, nil
-		}
-
-		getPlan := func(sel *sqlparser.Select) (logicalPlan, error) {
-			return newBuildSelectPlan(sel, reservedVars, vschema)
-		}
-
-		plan, err := getPlan(sel)
-		if err != nil {
-			return nil, err
-		}
-
-		if shouldRetryWithCNFRewriting(plan) {
-			// by transforming the predicates to CNF, the planner will sometimes find better plans
-			primitive := rewriteToCNFAndReplan(stmt, getPlan)
-			if primitive != nil {
-				return primitive, nil
-			}
-		}
-		return plan.Primitive(), nil
-	}
-}
-
-type postProcessor struct {
-	inDerived   bool
-	semTable    *semantics.SemTable
-	vschema     ContextVSchema
-	sqToReplace map[string]*sqlparser.Select
-}
-
-func (pp *postProcessor) planHorizon(plan logicalPlan, sel *sqlparser.Select, sqToReplace map[string]*sqlparser.Select) (logicalPlan, error) {
-	hp := horizonPlanning{
-		sel:       sel,
-		plan:      plan,
-		semTable:  pp.semTable,
-		vschema:   pp.vschema,
-		inDerived: pp.inDerived,
-	}
-
-	replaceSubQuery(sqToReplace, hp.sel)
-
-	plan, err := hp.planHorizon()
-	if err != nil {
-		return nil, err
-	}
-
-	plan, err = planLimit(sel.Limit, plan)
-	if err != nil {
-		return nil, err
-	}
-	return plan, nil
-
-}
-
-func newBuildSelectPlan(sel *sqlparser.Select, reservedVars *sqlparser.ReservedVars, vschema ContextVSchema) (logicalPlan, error) {
-	ksName := ""
-	if ks, _ := vschema.DefaultKeyspace(); ks != nil {
-		ksName = ks.Name
-	}
-	semTable, err := semantics.Analyze(sel, ksName, vschema, starRewrite)
-	if err != nil {
-		return nil, err
-	}
-
-	err = subqueryRewrite(sel, semTable, reservedVars)
-	if err != nil {
-		return nil, err
-	}
-
-	opTree, err := abstract.CreateOperatorFromSelect(sel, semTable)
-	if err != nil {
-		return nil, err
-	}
-
-	ctx := optimizeContext{
-		reservedVars: reservedVars,
-		semTable:     semTable,
-		vschema:      vschema,
-		sqToReplace:  map[string]*sqlparser.Select{},
-	}
-	tree, err := optimizeQuery(ctx, opTree)
-	if err != nil {
-		return nil, err
-	}
-
-	postProcessing := &postProcessor{
-		semTable:    semTable,
-		vschema:     vschema,
-		sqToReplace: ctx.sqToReplace,
-	}
-	plan, err := transformToLogicalPlan(tree, semTable, postProcessing)
-	if err != nil {
-		return nil, err
-	}
-
-	plan, err = postProcessing.planHorizon(plan, sel, postProcessing.sqToReplace)
-	if err != nil {
-		return nil, err
-	}
-
-	if err := setMiscFunc(plan, sel); err != nil {
-		return nil, err
-	}
-
-	if err := plan.WireupGen4(semTable); err != nil {
-		return nil, err
-	}
-
-	directives := sqlparser.ExtractCommentDirectives(sel.Comments)
-	if directives.IsSet(sqlparser.DirectiveScatterErrorsAsWarnings) {
-		visit(plan, func(logicalPlan logicalPlan) (bool, logicalPlan, error) {
-			switch plan := logicalPlan.(type) {
-			case *route:
-				plan.eroute.ScatterErrorsAsWarnings = true
-			}
-			return true, logicalPlan, nil
-		})
-	}
-
-	return plan, nil
-}
-
 type optimizeContext struct {
 	reservedVars *sqlparser.ReservedVars
 	semTable     *semantics.SemTable
@@ -295,132 +158,6 @@ func mergeSubQuery(ctx optimizeContext, outer *routeTree, subq *abstract.SubQuer
 	return outer, nil
 }
 
-func planLimit(limit *sqlparser.Limit, plan logicalPlan) (logicalPlan, error) {
-	if limit == nil {
-		return plan, nil
-	}
-	rb, ok := plan.(*route)
-	if ok && rb.isSingleShard() {
-		rb.SetLimit(limit)
-		return plan, nil
-	}
-
-	lPlan, err := createLimit(plan, limit)
-	if err != nil {
-		return nil, err
-	}
-
-	// visit does not modify the plan.
-	_, err = visit(lPlan, setUpperLimit)
-	if err != nil {
-		return nil, err
-	}
-	return lPlan, nil
-}
-
-type horizonPlanning struct {
-	sel             *sqlparser.Select
-	plan            logicalPlan
-	semTable        *semantics.SemTable
-	vschema         ContextVSchema
-	qp              *abstract.QueryProjection
-	inDerived       bool
-	needsTruncation bool
-	vtgateGrouping  bool
-}
-
-func (hp *horizonPlanning) planHorizon() (logicalPlan, error) {
-	rb, ok := hp.plan.(*route)
-	if !ok && hp.semTable.ProjectionErr != nil {
-		return nil, hp.semTable.ProjectionErr
-	}
-
-	if ok && rb.isSingleShard() {
-		createSingleShardRoutePlan(hp.sel, rb)
-		return hp.plan, nil
-	}
-
-	qp2, err := abstract.CreateQPFromSelect(hp.sel, hp.semTable)
-	if err != nil {
-		return nil, err
-	}
-
-	hp.qp = qp2
-
-	if err := checkUnsupportedConstructs(hp.sel); err != nil {
-		return nil, err
-	}
-
-	if hp.qp.NeedsAggregation() {
-		err = hp.planAggregations()
-		if err != nil {
-			return nil, err
-		}
-	} else {
-		_, isOA := hp.plan.(*orderedAggregate)
-		if isOA {
-			hp.plan = &simpleProjection{
-				logicalPlanCommon: newBuilderCommon(hp.plan),
-				primitive:         &engine.SimpleProjection{},
-			}
-		}
-		for _, e := range hp.qp.SelectExprs {
-			if _, _, err := pushProjection(e.Col, hp.plan, hp.semTable, true, false); err != nil {
-				return nil, err
-			}
-		}
-	}
-
-	if len(hp.qp.OrderExprs) > 0 {
-		hp.plan, err = hp.planOrderBy(hp.qp.OrderExprs, hp.plan)
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	if hp.qp.CanPushDownSorting && hp.vtgateGrouping {
-		hp.plan, err = hp.planOrderByUsingGroupBy()
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	err = hp.planDistinct()
-	if err != nil {
-		return nil, err
-	}
-
-	err = hp.truncateColumnsIfNeeded(hp.plan)
-	if err != nil {
-		return nil, err
-	}
-
-	return hp.plan, nil
-}
-
-func (hp *horizonPlanning) truncateColumnsIfNeeded(plan logicalPlan) error {
-	if !hp.needsTruncation {
-		return nil
-	}
-
-	switch p := plan.(type) {
-	case *route:
-		p.eroute.SetTruncateColumnCount(hp.sel.GetColumnCount())
-	case *joinGen4:
-		// since this is a join, we can safely add extra columns and not need to truncate them
-	case *orderedAggregate:
-		p.eaggr.SetTruncateColumnCount(hp.sel.GetColumnCount())
-	case *memorySort:
-		p.truncater.SetTruncateColumnCount(hp.sel.GetColumnCount())
-	case *pulloutSubquery:
-		return hp.truncateColumnsIfNeeded(p.underlying)
-	default:
-		return vterrors.Errorf(vtrpcpb.Code_INTERNAL, "plan type not known for column truncation: %T", hp.plan)
-	}
-
-	return nil
-}
-
 func exprHasUniqueVindex(vschema ContextVSchema, semTable *semantics.SemTable, expr sqlparser.Expr) bool {
 	col, isCol := expr.(*sqlparser.ColName)
 	if !isCol {
@@ -462,13 +199,6 @@ func createSingleShardRoutePlan(sel *sqlparser.Select, rb *route) {
 			ast.SelectExprs[i] = removeQualifierFromColName(aliasedExpr)
 		}
 	}
-}
-
-func checkUnsupportedConstructs(sel *sqlparser.Select) error {
-	if sel.Having != nil {
-		return semantics.Gen4NotSupportedF("HAVING")
-	}
-	return nil
 }
 
 func pushJoinPredicate(ctx optimizeContext, exprs []sqlparser.Expr, tree queryTree) (queryTree, error) {
